pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=68", "wheel"]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*.html", "*.css", "*.js", "*.ts", "*.json", "*.yaml", "*.yml"]

[project]
name = "nf-meta"
version = "0.1.1"
description = "nf-meta is a cli tool for building, validating and running meta-pipelines based on the Nextflow workflow language"
readme = "README.md"
authors = [
    { name = "Julian Flesch", email = "julian.flesch@qbic.uni-tuebingen.de" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "packaging>=26.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "uvicorn>=0.40.0",
]
license = "Apache-2.0"

[project.urls]
Homepage = "https://github.com/bmds-tue/nf-meta"
Issues = "https://github.com/bmds-tue/nf-meta/issues"

[project.scripts]
nf-meta = "nf_meta.__main__:cli"

[dependency-groups]
dev = [
    "twine>=6.2.0",
]
